docker/requirements.txt
tenacity
aiohttp
requests
apache-airflow-providers-databricks
azure-servicebus
databricks-sdk